[REFACTOR] 패키지 정리 -> aop 제거
