+++ build.gradle.kts
@@ -63,8 +63,7 @@ dependencies {
 
     // Cache
     implementation("org.springframework.boot:spring-boot-starter-cache")
-
-    implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.19.0")
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.3")
 
     // Test
     testImplementation("org.springframework.boot:spring-boot-starter-test")
